figuring out problems with syntax

+++ src/planners/bidirectional.jl
@@ -18,7 +18,7 @@ BiGreedyPlanner(f_heuristic::Heuristic, b_heuristic::Heuristic;  kwargs...) =
     )
 
 
-BidirectionalPlanner(f_heuristic::Heuristic, b_heuristic::Heuristic; max_nodes = typemax(Int), max_time = Inf, save_search = false) = 
+BidirectionalPlanner(f_heuristic::Heuristic, b_heuristic::Heuristic; max_nodes = typemax(Int64), max_time = Inf, save_search = false) = 
     BidirectionalPlanner(
         ForwardPlanner(;heuristic=f_heuristic, max_nodes, max_time, save_search),
         BackwardPlanner(;heuristic=b_heuristic, max_nodes, max_time, save_search),
